+++ go-controller/pkg/node/default_node_network_controller.go
@@ -1041,6 +1041,7 @@ func (nc *DefaultNodeNetworkController) Start(ctx context.Context) error {
 			nc.watchFactory.NodeInformer(),
 			nc.watchFactory.LocalPodInformer(),
 			informer.NewDefaultEventHandler,
+			false,
 		)
 		if err != nil {
 			return err
